Crate ic_replicated_state

Re-exports

pub use canister_state::num_bytes_from;

pub use canister_state::system_state::CanisterMetrics;

pub use canister_state::system_state::CanisterStatus;

pub use canister_state::system_state::SystemState;

pub use canister_state::CanisterState;

pub use canister_state::EmbedderCache;

pub use canister_state::ExecutionState;

pub use canister_state::ExportedFunctions;

pub use canister_state::Global;

pub use canister_state::NumWasmPages;

pub use canister_state::SchedulerState;

pub use metadata_state::NetworkTopology;

pub use metadata_state::NodeTopology;

pub use metadata_state::Stream;

pub use metadata_state::SubnetTopology;

pub use metadata_state::SystemMetadata;

pub use page_map::PageDelta;

pub use page_map::PageIndex;

pub use page_map::PageMap;

pub use replicated_state::ReplicatedState;

pub use replicated_state::StateError;

pub use stable_memory::StableMemory;

pub use stable_memory::StableMemoryError;

Modules

canister_state

metadata_state

page_map

replicated_state

stable_memory

testing

Structs

CallContext
Call context contains all context information related to an incoming call.

CallContextManager
Call context manager is an entity responsible for management of contexts of incoming calls of a canister. The call context manager must be used for opening new call contexts, registering and unregistering of callback for subsequent outgoing calls and for closing call contexts.

CanisterQueues
Wrapper around the induction pool (ingress and input queues); a priority queue used for round-robin scheduling of senders when consuming input messages; and output queues.

CyclesAccount
Cycles account

Enums

CallContextAction
The action the caller of CallContext.on_canister_result should take.

CallOrigin

CyclesAccountError