Module ic_replicated_state::canister_state

Re-exports

pub use execution_state::EmbedderCache;

pub use execution_state::ExecutionState;

pub use execution_state::ExportedFunctions;

pub use execution_state::Global;

Modules

execution_state

system_state

testing

Structs

CanisterQueues
Wrapper around the induction pool (ingress and input queues); a priority queue used for round-robin scheduling of senders when consuming input messages; and output queues.

CanisterState
The full state of a single canister.

NumWasmPagesTag

SchedulerState
State maintained by the scheduler.

Constants

QUEUE_INDEX_NONE
"None" queue index used internally by Message Routing for reject responses generated e.g. when a request cannot be inducted due to a full input queue (and enqueuing the response into the output queue might also fail).

Functions

num_bytes_from

Type Definitions

NumWasmPages
Count of number of Wasm Pages (which can be of different size than host page).

SessionNonce
A session is represented by an array of bytes and a monotonic offset and is unique for each execution.