Struct ic_replicated_state::canister_state::CanisterQueues

pub struct CanisterQueues { /* fields omitted */ }

Wrapper around the induction pool (ingress and input queues); a priority queue used for round-robin scheduling of senders when consuming input messages; and output queues.

Responsible for queue lifetime management, fair scheduling of inputs across sender canisters and queue backpressure.

Encapsulates the InductionPool component described in the spec. The reason for bundling together the induction pool and output queues is to reliably implement backpressure via queue reservations for response messages.

Implementations

impl CanisterQueues

pub fn push_ingress(&mut self, msg: Ingress)

Pushes an ingress message into the induction pool.

pub fn filter_ingress_messages<F>(&mut self, filter: F) where
F: FnMut(&Arc<Ingress>) -> bool,

See IngressQueue::filter_messages() for documentation

pub fn push_input(
&mut self,
index: QueueIndex,
msg: RequestOrResponse
) -> Result<(), (StateError, RequestOrResponse)>

Pushes a canister-to-canister message into the induction pool.

If the message is a Request this will also reserve a slot in the corresponding output queue for the eventual response.

If the message is a Response the protocol will have already reserved space for it, so the push cannot fail due to the input queue being full.

Errors

Returns a StateError along with the provided message:

QueueFull if pushing a Request and the corresponding input or output queues are full.

QueueFull if pushing a Response and the receiving canister is not expecting one.

pub fn has_input(&self) -> bool

Returns true if ingress_queue or at least one of the input_queues is not empty; false otherwise.

pub fn has_output(&self) -> bool

Returns true if at least one output queue is not empty; false otherwise.

pub fn pop_input(&mut self) -> Option<CanisterInputMessage>

Extracts the next inter-canister or ingress message (in that order). If no inter-canister messages are available in the induction pool, we pop the next ingress message.

pub fn push_output_request(
&mut self,
msg: Request
) -> Result<(), (StateError, Request)>

Pushes a Request type message into the relevant output queue. Also reserves a slot for the eventual response on the matching input queue.

Errors

Returns a QueueFull error along with the provided message if either the output queue or the matching input queue is full.

pub fn push_output_response(&mut self, msg: Response)

Pushes a Response type message into the relevant output queue. The protocol should have already reserved a slot, so this cannot fail.

Panics

Panics if the queue does not already exist or there is no reserved slot to push the Response into.

pub fn output_into_iter<'a>(
&'a mut self,
owner: CanisterId
) -> impl Iterator<Item = (QueueId, QueueIndex, RequestOrResponse)> + 'a

Returns an iterator that consumes all output messages.