Struct ic_replicated_state::canister_state::CanisterState

pub struct CanisterState {
    pub system_state: SystemState,
    pub execution_state: Option<ExecutionState>,
    pub scheduler_state: SchedulerState,
}

The full state of a single canister.

Fields
system_state: SystemState

See SystemState for documentation.

execution_state: Option<ExecutionState>

See ExecutionState for documentation.

This may or may not exist depending on whether or not the canister has an actual wasm module. A valid canister is not required to contain a Wasm module. Canisters without Wasm modules can exist as a store of DFNs; temporarily when they are being upgraded, etc.

scheduler_state: SchedulerState

See SchedulerState for documentation.

Implementations

impl CanisterState

pub fn new(
system_state: SystemState,
execution_state: Option<ExecutionState>,
scheduler_state: SchedulerState
) -> Self

pub fn canister_id(&self) -> CanisterId

pub fn controller(&self) -> PrincipalId

pub fn push_input(
&mut self,
index: QueueIndex,
msg: RequestOrResponse
) -> Result<(), (StateError, RequestOrResponse)>

See SystemState::push_input for documentation.

Warning! this is only exposed for test purposes and should not be used outside of this crate.

pub fn pop_input(&mut self) -> Option<CanisterInputMessage>

See SystemState::pop_input for documentation.

pub fn has_input(&self) -> bool

See SystemState::has_input for documentation.

pub fn has_output(&self) -> bool

Returns true if there is at least one message in the canister's output queues, false otherwise.

pub fn push_output_request(
&mut self,
msg: Request
) -> Result<(), (StateError, Request)>

See SystemState::push_output_request for documentation.

pub fn push_output_response(&mut self, msg: Response)

See SystemState::push_output_response for documentation.

pub fn push_ingress(&mut self, msg: Ingress)

Unconditionally pushes an ingress message into the input queue of the canister.

pub fn output_into_iter<'a>(
&'a mut self
) -> impl Iterator<Item = (QueueId, QueueIndex, RequestOrResponse)> + 'a

See CanisterQueues::output_into_iter for documentation.

pub fn into_parts(self) -> (Option<ExecutionState>, SystemState, SchedulerState)

pub fn from_parts(
execution_state: Option<ExecutionState>,
system_state: SystemState,
scheduler_state: SchedulerState
) -> Self

pub fn memory_usage(&self) -> NumBytes

The amount of memory currently being used by the canister.

pub fn memory_allocation(&self) -> Option<NumBytes>

Returns the current memory allocation of the canister in bytes.

pub fn compute_allocation(&self) -> ComputeAllocation

Returns the current compute allocation for the canister.

pub fn exports_update_method(&self, method_name: String) -> bool

Returns true if the canister contains an exported update method with the name provided, false otherwise.

pub fn exports_query_method(&self, method_name: String) -> bool

Returns true if the canister contains an exported query method with the name provided, false otherwise.

pub fn exports_system_method(&self, method: SystemMethod) -> bool

Returns true if the canister contains an exported system method with the name provided, false otherwise.

pub fn num_wasm_globals(&self) -> usize

Returns the number of global variables in the Wasm module.

pub fn status(&self) -> CanisterStatusType