Module ic_replicated_state::canister_state::system_state

Structs

CallContext
Call context contains all context information related to an incoming call.

CallContextManager
Call context manager is an entity responsible for management of contexts of incoming calls of a canister. The call context manager must be used for opening new call contexts, registering and unregistering of callback for subsequent outgoing calls and for closing call contexts.

CanisterMetrics
Canister-specific metrics on scheduling, maintained by the scheduler.

CyclesAccount
Cycles account

SystemState
State that is controlled and owned by the system (IC).

Enums

CallContextAction
The action the caller of CallContext.on_canister_result should take.

CallOrigin

CanisterStatus
A wrapper around the different canister statuses.

CyclesAccountError