Module ic_replicated_state::metadata_state

Structs

IngressHistoryState
State associated with the history of statuses of ingress messages as they traversed through the system.

NetworkTopology

NodeTopology

Stream
Stream is the state of bi-directional communication session with a remote subnet. It contains outgoing messages having that subnet as their destination and signals for inducted messages received from that subnet.

SubnetCallContextManager

SubnetTopology

SystemMetadata
Replicated system metadata. Used primarily for inter-canister messaging and history queries.

Enums

SubnetCallContext

Type Definitions

Streams