Struct ic_replicated_state::page_map::PageDelta

pub struct PageDelta(_);

PageDelta represents a changeset of the module heap.

Implementations

impl PageDelta

pub fn is_empty(&self) -> bool

Returns true if the delta doesn't contain any pages.

pub fn len(&self) -> usize

Returns the number of pages in the delta.

pub fn get_page(&self, page_index: PageIndex) -> Option<&[u8]>

Gets content of the page at the specified index.

pub fn update(&mut self, rhs: PageDelta)

Modifies this delta in-place by applying all the entries in rhs to it.

pub fn extend(self, rhs: PageDelta) -> Self

Unions two page deltas into a single delta. If both deltas contain a page with the same number, the page contained in self wins, i.e. extend([(N, X)], [(N, Y), (M, Z)]) = [(N, X), (M, Z)]

pub fn iter(&self) -> impl Iterator<Item = (PageIndex, &TrackedPage)>

Enumerates all the pages in this delta.

pub fn persist(&self, dst: &Path) -> Result<(), PersistenceError>

Persists this delta to the specified destination.

pub fn persist_and_flush(&self, dst: &Path) -> Result<(), PersistenceError>

Persists this delta to the specified destination and flushes it.