Struct ic_replicated_state::replicated_state::ReplicatedState

pub struct ReplicatedState {
    pub canister_states: BTreeMap<CanisterId, CanisterState>,
    pub metadata: SystemMetadata,
    pub subnet_queues: CanisterQueues,
    pub consensus_queue: Vec<Response>,
    pub root: PathBuf,
}

ReplicatedState is the deterministic replicated state of the system. Broadly speaking it consists of two parts: CanisterState used for canister execution and SystemMetadata used for message routing and history queries.

Fields
canister_states: BTreeMap<CanisterId, CanisterState>

States of all canisters, indexed by canister ids.

metadata: SystemMetadata

Deterministic processing metadata.

subnet_queues: CanisterQueues

Queues for holding messages sent/received by the subnet.

consensus_queue: Vec<Response>

Queue for holding responses arriving from Consensus.

Responses from consensus are to be processed each round. The queue is, therefore, emptied at the end of every round.

root: PathBuf
Implementations

impl ReplicatedState

pub fn new_rooted_at(
own_subnet_id: SubnetId,
own_subnet_type: SubnetType,
root: PathBuf
) -> ReplicatedState

Creates a new empty node state.

pub fn path(&self) -> &Path

pub fn canister_state(&self, canister_id: &CanisterId) -> Option<&CanisterState>

pub fn canister_state_mut(
&mut self,
canister_id: &CanisterId
) -> Option<&mut CanisterState>

pub fn take_canister_state(
&mut self,
canister_id: &CanisterId
) -> Option<CanisterState>

pub fn take_canister_states(&mut self) -> BTreeMap<CanisterId, CanisterState>

pub fn put_canister_state(&mut self, canister_state: CanisterState)

Insert the canister state into the replicated state. If a canister already exists for the given canister id, it will be replaced. It is the responsibility of the caller of this function to ensure that any relevant state associated with the older canister state are properly cleaned up.

pub fn put_canister_states(
&mut self,
canisters: BTreeMap<CanisterId, CanisterState>
)

pub fn canisters_iter(&self) -> Values<CanisterId, CanisterState>

Returns an iterator over canister states, ordered by canister ID.

pub fn canisters_iter_mut(&mut self) -> ValuesMut<CanisterId, CanisterState>

Returns a mutable iterator over canister states, ordered by canister ID.

pub fn system_metadata(&self) -> &SystemMetadata

pub fn set_system_metadata(&mut self, metadata: SystemMetadata)

pub fn get_ingress_status(&self, message_id: &MessageId) -> IngressStatus

pub fn get_ingress_history(&self) -> IngressHistoryState

pub fn set_ingress_status(
&mut self,
message_id: MessageId,
status: IngressStatus
)

pub fn prune_ingress_history(&mut self)

pub fn take_streams(&mut self) -> Streams

Removes the Streams from this ReplicatedState.

pub fn put_streams(&mut self, streams: Streams)

Atomically updates streams to the provided ones.

pub fn streams(&self) -> &Streams

Returns a reference to all streams.

pub fn subnets_with_available_streams(&self) -> Vec<SubnetId>

Returns all subnets for which a stream is available.

pub fn get_stream(&self, destination_subnet_id: SubnetId) -> Option<&Stream>

Retrieves a reference to the stream from this subnet to the destination subnet, if such a stream exists.

pub fn get_mut_stream(
&mut self,
destination_subnet_id: SubnetId
) -> Option<&mut Stream>

Retrieves a mutable reference to the stream from this subnet to the destination subnet, if such a stream exists.

pub fn modify_streams<F: FnOnce(&mut Streams)>(&mut self, f: F)

pub fn total_compute_allocation(&self) -> u64

Returns the sum of reserved compute allocations of all currently available canisters.

pub fn total_memory_allocation_used(&self) -> NumBytes

pub fn push_input(
&mut self,
index: QueueIndex,
msg: RequestOrResponse
) -> Result<(), (StateError, RequestOrResponse)>

Pushes a RequestOrResponse into the induction pool. The induction pool can either be that of a canister or that of the subnet.

pub fn time(&self) -> Time

pub fn induct_messages_on_same_subnet(&mut self, log: &ReplicaLogger)

Iterates over all canisters on the subnet, checking if a source canister has output messages for a destination canister on the same subnet and moving them from the source to the destination canister if the destination canister has room for them.

This method only handles messages sent to actual other canisters. Messages sent to the subnet or self are not handled i.e. they take the slow path through message routing.