Trait ic_replicated_state::testing::CanisterQueuesTesting

pub trait CanisterQueuesTesting {
    fn ingress_queue_size(&self) -> usize;

    fn pop_canister_output(
        &mut self, 
        dst_canister: &CanisterId
    ) -> Option<(QueueIndex, RequestOrResponse)>;

    fn output_queues_len(&self) -> usize;

    fn output_message_count(&self) -> usize;
}

Exposes public testing-only CanisterQueues methods to be used in other crates' unit tests.

Required methods

fn ingress_queue_size(&self) -> usize

Returns the number of messages in ingress_queue.

fn pop_canister_output(
&mut self,
dst_canister: &CanisterId
) -> Option<(QueueIndex, RequestOrResponse)>

Pops the next message from the output queue associated with dst_canister. Returned QueueIndex values are sequential across successful calls.

fn output_queues_len(&self) -> usize

Returns the number of output queues, empty or not.

fn output_message_count(&self) -> usize

Returns the total number of messages in the output queues.

Loading content...